# use sum tree for O(log(n)) searching