This is a faster prioritised learning search alg as suggested in the original paper.

Speed:

|search space|old speed (ms)|numpy (ms)| sum-tree (ms) |
|-|-|-|-|
|1e3|13.9 |0.9| 1.45 |
|1e4|132.0 |4.7| 1.7 |
|1e5|1690.0 |55.6| 2.1 |
|Order|O(n**1.04)|O(n**0.9)|~O(log(n)) or O(n**0.08)|