@michaelschaarschmidt 2017-10-22
Hey,

we are looking to merge this now. Some comments:
- We want to implement the memories in TensorFlow as well and it is not clear whether this strategy would be easily implemented there
- One way would be to keep the old prioritised replay and have this one as a separate class first so we can have this version in python if it's inconvenient in TF