@wassname 2017-10-22
Good points, and in the end you decide and I'll go with that.

My thoughts are that the old one will be kept around in the older commits, that way it can be used as reference to make the tensorflow version. This implementation might be able to be ported to tensorflow (it revolved around a list of None's and floats, but is not very vectorized). And at that point the tensorflow version would be the best.

But if you decide to keep both for that reason, it's fine with me. Want me to implement that, and fix the conflicts?