@AlexKuhnle 2017-10-22
I think it would be best if for now you could rename the old version to `naive_prioritized_replay` and yours as the 'default' `prioritized_replay`. We can then see in the next weeks whether we get rid of it. If you could do that, would be great. :+1: