@wassname 2017-10-22
Sounds good.

And thanks for all the work on 0.3.0, it's awesome!