@michaelschaarschmidt 2017-10-22
Thanks for implementing!